refactor: rename components dir
